+++ src/repositorios/mongo/implementacoes/municipios/TruncateMunicipiosMongo.ts
@@ -1,11 +1,10 @@
 import { ClienteMongo } from '../../ClienteMongo'
 import { IRepositorio } from '../../../IRepositorio'
-import { AbastractRepositorioMunicipios } from '../../../AbastractRepositorioMunicipios'
+import { RepositorioMunicipios } from '../../../RepositorioMunicipios'
 
-class TruncateMunicipiosMongo extends AbastractRepositorioMunicipios implements IRepositorio<void> {
+class TruncateMunicipiosMongo implements IRepositorio<void> {
   public queryObj: { }
   constructor () {
-    super()
     this.queryObj = { }
   }
 
@@ -14,12 +13,25 @@ class TruncateMunicipiosMongo extends AbastractRepositorioMunicipios implements 
     try {
       await clienteMongo.connect()
 
-      await clienteMongo
+      const collections = await clienteMongo
         .db(process.env.MONGO_INIT_DB)
-        .collection(this.SCHEMA_NAME)
-        .drop()
-    } finally {
+        .listCollections()
+        .toArray()
+
+      collections.forEach(async c => {
+        if (c.name === RepositorioMunicipios.SCHEMA_NAME) {
+          const clienteMongo2 = (new ClienteMongo()).client
+          clienteMongo2.connect()
+          await clienteMongo2
+            .db(process.env.MONGO_INIT_DB)
+            .collection(RepositorioMunicipios.SCHEMA_NAME)
+            .drop()
+          clienteMongo2.close()
+        }
+      })
       await clienteMongo.close()
+    } catch (err) {
+      console.error(err)
     }
   }
 }
